example/requirements.txt
grpcio==1.73.0
grpcio-tools==1.73.0
pandas==2.2.2
